8/8 tests so far
